Implement initial location by geo ip
Initial map location is set according to a geo ip lookup (if feature
is enabled). Allows the map to start roughly at the user's location
in the case that location services are disabled.

+++ src/main/java/ldprotest/db/codec/BsonSerializableCodec.java
@@ -512,7 +512,7 @@ public class BsonSerializableCodec <T> implements Codec<T> {
         );
         map.put(
             Double.TYPE, (reader, field, object) -> {
-                field.setDouble(object, (short)reader.readInt32());
+                field.setDouble(object, (short)reader.readDouble());
             }
         );
         map.put(
